refactor: mover tests de dentadura a tests/Feature/Clientes/
